# project-pipeline/backend/openapi.yaml
openapi: 3.0.0
info:
  title: Project Pipeline Management API
  description: |
    AI-powered Erasmus+ project planning and management system.

    Features:
    - 6 specialized AI agents for different planning domains
    - Real-time budget tracking and alerts
    - Automated vendor communications
    - Learning system that improves with each project
    - Erasmus+-compliant reporting

  version: 1.0.0
  contact:
    name: API Support
    email: support@projectpipeline.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.projectpipeline.com
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [STUDENT_EXCHANGE, TEACHER_TRAINING, VOCATIONAL, YOUTH_EXCHANGE, OTHER]
        description:
          type: string
        location:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        participants:
          type: integer
        budget:
          type: number
          format: float
        status:
          type: string
          enum: [PLANNING, IN_PROGRESS, COMPLETED, CANCELLED]
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: List all projects
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
    post:
      tags: [Projects]
      summary: Create new project
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, type, startDate, endDate, participants, budget]
              properties:
                name:
                  type: string
                type:
                  type: string
                description:
                  type: string
                location:
                  type: string
                startDate:
                  type: string
                  format: date
                endDate:
                  type: string
                  format: date
                participants:
                  type: integer
                budget:
                  type: number
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
  /projects/{id}:
    get:
      tags: [Projects]
      summary: Get project details
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    patch:
      tags: [Projects]
      summary: Update project
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Project updated
    delete:
      tags: [Projects]
      summary: Delete project
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Project deleted

  # Budget
...